+++ index.html
@@ -4,15 +4,17 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Chess Game Made By TK</title>
-        <link rel="stylesheet" href="./src/board/board.css">
+        <link rel="stylesheet" href="./src/board/chessboard.css">
         <link rel="stylesheet" href="./dist/output.css">
 
         <script defer type="module" src="./index.js"></script>
     </head>
     <body id="body">
+        <!-- Pawn promotion modal -->
         <div id="popup-container">
         </div>
 
+        <!-- Game has finished Modal -->
         <div id="game-finished" class="hidden absolute flex flex-col space-around h-3/5 w-2/5 rounded-2xl bg-amber-900">
             <div id="winner-section" class="flex flex-col justify-evenly h-1/3 w-full items-center justify-center">
                 <h1 id="checkmate" class="text-4xl font-semibold text-white">CHECKMATE!!!</h1>
@@ -20,11 +22,11 @@
             </div>
 
             <div id="options" class="flex w-full p-4 h-3/5 items-center justify-evenly">
-
+                <!-- Reset game back to the beginning -->
                 <div id="reset" class="flex items-center justify-center border-2 h-32 w-5/12 rounded-2xl bg-red-300">
                     <p class="text-2xl font-semibold text-white">RESET</p>
                 </div>
-
+                <!-- Leave the game as is -->
                 <div id="stay" class="flex items-center justify-center border-2 h-32 w-5/12 rounded-2xl bg-green-300">
                     <p class="text-2xl font-semibold text-white">STAY</p>
                 </div>
