+++ index.js
@@ -1,4 +1,4 @@
-import { createBoard, renderChessboard, reRenderChessboard } from "./src/board/board.js";
+import { createChessboard, initialChessboardRender, reRenderChessboard } from "./src/board/chessboard.js";
 import { initializeBoardWithPieces } from "./src/pieces/initialise-pieces.js";
 import { removeAllHighlightClasses, highlightTileOnly, pressedTile } from "./src/util/clickedPiece.js";
 import { movePiece, pieceOrTile, tileFullLocation } from "./src/util/movement/movePiece.js";
@@ -7,24 +7,26 @@ import { clickedEnemyPiece} from "./src/util/management/gameManagement.js";
 import { checkmate } from "./src/util/checkmate/checkmate.js";
 import { gameHasEnded } from "./src/util/management/managementDesign.js";
  
-export const chessBoard = createBoard();
-
+// Create and render the chessboard
+export const chessBoard = createChessboard();
 initializeBoardWithPieces(chessBoard)
+initialChessboardRender(chessBoard);
 
-renderChessboard(chessBoard);
-
+// Background and Chessboard element
 const bodyElement = document.getElementById('body');
 const chessBoardElement = document.querySelector('.chess-board');
 
+// Track initial selected piece
 let initialSelectedPiece;
 let initialSelectedPieceLocation;
 
-// Keep track of kings for check / checkmate
+// Track Kings
 const KINGS = {
     "white": chessBoard[7][4].spaceOccupation,
     "black": chessBoard[0][4].spaceOccupation
 }
 
+// When user presses off the chessboard
 bodyElement.addEventListener('click', (event) => {
     const isClickedOnChessboard = chessBoardElement.contains(event.target);
 
@@ -34,14 +36,18 @@ bodyElement.addEventListener('click', (event) => {
     }
 });
 
+// Deals with when a tile has been pressed
 const tilePressed = async (tileCheck, event) => {
+    // User presses an empty tile
     if (tileCheck && !initialSelectedPiece) {
         highlightTileOnly(event, chessBoard);
+    // User initially presses on a piece
     } else if (!tileCheck && !initialSelectedPiece) {
         initialSelectedPieceLocation = tileFullLocation(event);
         initialSelectedPiece = await pressedTile(event, chessBoard);
 
-        clickedEnemyPiece(initialSelectedPiece, event)
+        clickedEnemyPiece(initialSelectedPiece, event);
+    // User pressed on a piece and decides to move it
     } else {
         await movePiece(initialSelectedPiece, initialSelectedPieceLocation, event, chessBoard);
         reRenderChessboard(chessBoard)
@@ -52,6 +58,7 @@ const tilePressed = async (tileCheck, event) => {
     }
 }
 
+// When user presses on chessboard
 chessBoardElement.addEventListener('click', async (event) => {
     if (PLAYERGAME.status !== "finished") {
         const tileCheck = pieceOrTile(event);
